refactor(notification): ship inbound receivers from the provider subpaths
A channel is one unit, so surface both halves from the same entry: each
@visulima/notification/providers/<name> now exports the outbound provider AND
the inbound receiver (create<Name>Inbound + options). Drop the separate
./inbound/<name> package.json exports and stop re-exporting channels from the
./inbound barrel, which now holds only the shared types, createInboundRouter
and the verification/reply helpers.

Channel receiver code stays in src/inbound/channels/* and is bundled into each
provider entry (tree-shaken when unused). Docs, README and tests import each
receiver from its provider subpath.
