+++ packages/notification/notification/docs/inbound.mdx
@@ -20,7 +20,7 @@ standard `Request`/`Response`.
 ## Quick start
 
 ```typescript
-import { createSlackInbound } from "@visulima/notification/inbound/slack";
+import { createSlackInbound } from "@visulima/notification/providers/slack";
 
 const slack = createSlackInbound({
     signingSecret: process.env.SLACK_SIGNING_SECRET!,
@@ -84,8 +84,8 @@ response. For everything else (Slack Events API, Discord deferred replies), pass
 message to the right outbound payload (recipient, thread, WhatsApp prefix) and sends it:
 
 ```typescript
-import { createSlackInbound } from "@visulima/notification/inbound/slack";
-import slackProvider from "@visulima/notification/providers/slack";
+// The inbound receiver and the outbound provider ship from the same channel subpath.
+import { createSlackInbound, slackProvider } from "@visulima/notification/providers/slack";
 
 const slack = createSlackInbound({
     signingSecret,
@@ -101,16 +101,21 @@ const slack = createSlackInbound({
 
 ## Channels
 
-| Import                                    | Factory                  | Verification                                              | Synchronous reply                     |
-| ----------------------------------------- | ------------------------ | -------------------------------------------------------- | ------------------------------------- |
-| `@visulima/notification/inbound/slack`    | `createSlackInbound`     | v0 signature + 5-min replay; `url_verification` handled  | Slash commands & interactions         |
-| `@visulima/notification/inbound/discord`  | `createDiscordInbound`   | Ed25519 over `timestamp+body`; `PING` → `PONG` handled   | Interaction response                  |
-| `@visulima/notification/inbound/telegram` | `createTelegramInbound`  | `X-Telegram-Bot-Api-Secret-Token` (when set)             | `sendMessage` method in the response  |
-| `@visulima/notification/inbound/twilio`   | `createTwilioInbound`    | `X-Twilio-Signature` HMAC-SHA1 of URL + sorted params    | TwiML message element                 |
-| `@visulima/notification/inbound/msteams`  | `createMsTeamsInbound`      | `Authorization: HMAC` — base64 HMAC-SHA256 of raw body   | Bot Framework message activity     |
-| `@visulima/notification/inbound/telnyx`   | `createTelnyxInbound`       | Ed25519 over `{timestamp}\|{body}` + 5-min replay        | via `context.reply()` (async)      |
-| `@visulima/notification/inbound/messagebird` | `createMessageBirdInbound` | `messagebird-signature-jwt` — HS256 JWT + `url_hash`/`payload_hash` | via `context.reply()` (async) |
-| `@visulima/notification/inbound/vonage`   | `createVonageInbound`       | `Authorization: Bearer` HS256 JWT + `payload_hash`       | via `context.reply()` (async)      |
+A channel's two halves ship together: each `@visulima/notification/providers/<name>` subpath
+exports both the outbound provider (`<name>Provider`) and the inbound receiver
+(`create<Name>Inbound`). The `@visulima/notification/inbound` barrel holds only the shared types,
+`createInboundRouter` and the verification/reply helpers.
+
+| Import                                     | Receiver                   | Verification                                              | Synchronous reply                  |
+| ------------------------------------------ | -------------------------- | -------------------------------------------------------- | ---------------------------------- |
+| `@visulima/notification/providers/slack`   | `createSlackInbound`       | v0 signature + 5-min replay; `url_verification` handled  | Slash commands & interactions      |
+| `@visulima/notification/providers/discord` | `createDiscordInbound`     | Ed25519 over `timestamp+body`; `PING` → `PONG` handled   | Interaction response               |
+| `@visulima/notification/providers/telegram`| `createTelegramInbound`    | `X-Telegram-Bot-Api-Secret-Token` (when set)             | `sendMessage` method               |
+| `@visulima/notification/providers/twilio`  | `createTwilioInbound`      | `X-Twilio-Signature` HMAC-SHA1 of URL + sorted params    | TwiML message element              |
+| `@visulima/notification/providers/msteams` | `createMsTeamsInbound`     | `Authorization: HMAC` — base64 HMAC-SHA256 of raw body   | Bot Framework message activity     |
+| `@visulima/notification/providers/telnyx`  | `createTelnyxInbound`      | Ed25519 over `{timestamp}\|{body}` + 5-min replay        | via `context.reply()` (async)      |
+| `@visulima/notification/providers/messagebird` | `createMessageBirdInbound` | `messagebird-signature-jwt` HS256 JWT + `url_hash`/`payload_hash` | via `context.reply()` (async) |
+| `@visulima/notification/providers/vonage`  | `createVonageInbound`      | `Authorization: Bearer` HS256 JWT + `payload_hash`       | via `context.reply()` (async)      |
 
 Slack receivers cover the Events API, slash commands and interactive components; Twilio, Telnyx,
 MessageBird and Vonage cover SMS (Twilio also WhatsApp — flagged in `metadata.transport`, with
@@ -136,8 +141,8 @@ implements the full RSA certificate-chain scheme (see the [webhooks docs](/docs/
 
 ```typescript
 import { createInboundRouter } from "@visulima/notification/inbound";
-import { createSlackInbound } from "@visulima/notification/inbound/slack";
-import { createTelegramInbound } from "@visulima/notification/inbound/telegram";
+import { createSlackInbound } from "@visulima/notification/providers/slack";
+import { createTelegramInbound } from "@visulima/notification/providers/telegram";
 
 const handler = createInboundRouter({
     "/webhooks/slack": createSlackInbound({ signingSecret, onMessage }),
@@ -153,8 +158,7 @@ Inbound + outbound together give you a receive → think → reply loop in a sin
 the outbound provider once and answer with `context.reply()`:
 
 ```typescript
-import { createSlackInbound } from "@visulima/notification/inbound/slack";
-import slackProvider from "@visulima/notification/providers/slack";
+import { createSlackInbound, slackProvider } from "@visulima/notification/providers/slack";
 
 const slack = createSlackInbound({
     signingSecret,
