+++ packages/notification/notification/package.json
@@ -196,38 +196,6 @@
             "types": "./dist/inbound/index.d.ts",
             "default": "./dist/inbound/index.js"
         },
-        "./inbound/slack": {
-            "types": "./dist/inbound/channels/slack.d.ts",
-            "default": "./dist/inbound/channels/slack.js"
-        },
-        "./inbound/discord": {
-            "types": "./dist/inbound/channels/discord.d.ts",
-            "default": "./dist/inbound/channels/discord.js"
-        },
-        "./inbound/telegram": {
-            "types": "./dist/inbound/channels/telegram.d.ts",
-            "default": "./dist/inbound/channels/telegram.js"
-        },
-        "./inbound/twilio": {
-            "types": "./dist/inbound/channels/twilio.d.ts",
-            "default": "./dist/inbound/channels/twilio.js"
-        },
-        "./inbound/msteams": {
-            "types": "./dist/inbound/channels/msteams.d.ts",
-            "default": "./dist/inbound/channels/msteams.js"
-        },
-        "./inbound/telnyx": {
-            "types": "./dist/inbound/channels/telnyx.d.ts",
-            "default": "./dist/inbound/channels/telnyx.js"
-        },
-        "./inbound/messagebird": {
-            "types": "./dist/inbound/channels/messagebird.d.ts",
-            "default": "./dist/inbound/channels/messagebird.js"
-        },
-        "./inbound/vonage": {
-            "types": "./dist/inbound/channels/vonage.d.ts",
-            "default": "./dist/inbound/channels/vonage.js"
-        },
         "./template/string": {
             "types": "./dist/template-engines/string.d.ts",
             "default": "./dist/template-engines/string.js"
