+++ packages/notification/notification/src/providers/chat/discord/index.ts
@@ -1,2 +1,4 @@
+export type { DiscordInboundOptions } from "../../../inbound/channels/discord";
+export { createDiscordInbound } from "../../../inbound/channels/discord";
 export { default as discordProvider } from "./provider";
 export type { DiscordConfig } from "./types";
