+++ packages/notification/notification/src/providers/chat/msteams/index.ts
@@ -1,2 +1,4 @@
+export type { MsTeamsInboundOptions } from "../../../inbound/channels/msteams";
+export { createMsTeamsInbound } from "../../../inbound/channels/msteams";
 export { default as msTeamsProvider } from "./provider";
 export type { MsTeamsConfig } from "./types";
