+++ packages/notification/notification/src/providers/chat/slack/index.ts
@@ -1,2 +1,4 @@
+export type { SlackInboundOptions } from "../../../inbound/channels/slack";
+export { createSlackInbound } from "../../../inbound/channels/slack";
 export { default as slackProvider } from "./provider";
 export type { SlackConfig } from "./types";
