+++ packages/notification/notification/src/providers/chat/telegram/index.ts
@@ -1,2 +1,4 @@
+export type { TelegramInboundOptions } from "../../../inbound/channels/telegram";
+export { createTelegramInbound } from "../../../inbound/channels/telegram";
 export { default as telegramProvider } from "./provider";
 export type { TelegramConfig } from "./types";
