+++ packages/notification/notification/src/providers/sms/messagebird/index.ts
@@ -1,2 +1,4 @@
+export type { MessageBirdInboundOptions } from "../../../inbound/channels/messagebird";
+export { createMessageBirdInbound } from "../../../inbound/channels/messagebird";
 export { default as messageBirdProvider } from "./provider";
 export type { MessageBirdConfig } from "./types";
