+++ packages/notification/notification/src/providers/sms/telnyx/index.ts
@@ -1,2 +1,4 @@
+export type { TelnyxInboundOptions } from "../../../inbound/channels/telnyx";
+export { createTelnyxInbound } from "../../../inbound/channels/telnyx";
 export { default as telnyxProvider } from "./provider";
 export type { TelnyxConfig } from "./types";
