+++ packages/notification/notification/src/providers/sms/twilio/index.ts
@@ -1,2 +1,4 @@
+export type { TwilioInboundOptions } from "../../../inbound/channels/twilio";
+export { createTwilioInbound } from "../../../inbound/channels/twilio";
 export { default as twilioProvider } from "./provider";
 export type { TwilioConfig } from "./types";
