+++ packages/notification/notification/src/providers/sms/vonage/index.ts
@@ -1,2 +1,4 @@
+export type { VonageInboundOptions } from "../../../inbound/channels/vonage";
+export { createVonageInbound } from "../../../inbound/channels/vonage";
 export { default as vonageProvider } from "./provider";
 export type { VonageConfig } from "./types";
